fix(io): fix unexpected MPI-IO behavior in copy output mode under OpenMPI
Here, the observational unexpected MPI-IO behavior is: the md-tools conversion tool does
not exit with a very large text file (data stored in the  MD output binary file is incorrect).
The behavior is cause by the incorrect extent, as well as upper/lower bounds, of the derived
MPI Datatype, which causes different size calculated by `MPI_Type_size` under MPICH and OpenMPI
(in MPICH, the conversion result is as our expected, but in OpenMP it does not.)

This can be easily fixed by adding `MPI_Type_create_resized`.

fix #21 on git.hpcer.dev