test(collision): correct the failed test `WorldBuilderTestFixture.atom_set_v_test`
The first parameter `lat` passed to `atom::set_v` is type of int[4], instead of int[3]. Before this
commit, we pass type of int[3] to `atom::set_v`, which can cause the 4th element in `lat` is an
unexpected value. We fix it by passing type of int[4]. Besides, another bug in the test code is also
fixed: loading atom using `getAtomIndexInSubBox` instead of `getAtomIndex`.